refactor(core): split response-contracts into output-schemas and formatters
