+++ .changeset/split-response-contracts.md
@@ -1,0 +1,8 @@
+---
+"@hevy-mcp/core": patch
+"hevy-mcp": patch
+"@hevy-mcp/worker": patch
+"@chrisdoc/hevy-cli": patch
+---
+
+Split response-contracts into output-schemas and formatters modules with no behavior change; response contract wiring stays in response-contracts.
